Merge pull request #48 from Capg-RCOEM-Group-1/mandar

+++ src/main/java/com/rcoem/filmrentalstore/entities/Actor.java
@@ -23,7 +23,7 @@ public class Actor {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long actorId;
+    private Short actorId;
 
     
     @Column(nullable = false)
